reorganizing for headless run setup
