+++ scripts/replay_stubs.py
