+++ src/participation_model.py
